Created the Initial Layout & Started Work
This block of code shows the map and also allows a user to
search for a specific location. In either case, it centers the
map for the user. Long press gesture allows the user to drop
map point. There are a variety of things left to do still
See todo list