Memorable places (#15)
* Created the Initial Layout & Started Work

This block of code shows the map and also allows a user to
search for a specific location. In either case, it centers the
map for the user. Long press gesture allows the user to drop
map point. There are a variety of things left to do still
See todo list

* Not alot of time today

Added the UIAlert to capture what the user would like to call their
favorite place. Need to add this to local storage, and then display
all local places to the tableview on the main view.

* Added the code to save and retrieve state from
userdefaults. However, the initial tableview
doesn't appear to be loading when the app is.

* Added Delete & Backing Class for the TableView

* App Connects all the pieces together

It includes maps and location. It includes state and it includes
segues.